requirements.txt
httpx[socks]>=0.27.0    # Async HTTP client with SOCKS5 proxy support
tenacity>=8.2.0         # Retry with exponential backoff
pillow>=10.0.0          # Image processing
img2pdf>=0.5.0          # PDF generation from images
click>=8.0.0            # CLI framework
python-dotenv>=1.0.0    # Environment variable loading
tqdm>=4.66.0            # Progress bar
